Initial architechture and functionality
This commit contains allot of the initial architechture for bonfire.js.

It's able to get the data from firebase and detect changes in the bonfire.

However, it's unable to react to changes in Firebase and sync them to the bonfire (which is a huge challenge).

It also doesn't know how to handle object sets to the bonfire object (so you can't do foo.bar = {a: {b:"baz"}} yet), simply because I didn't have time to implement that yet (it's just a matter of adding some code in the setProxyHandler to detect that the value of the change is an object, and calling bonify on it.

It also doesn't know how to handle firebase "arrays" (pushIDs etc), and has not been tested on arrays of any kind at all....